weight implementation for some 3

+++ src/main/java/com/megatrex4/InventoryWeightState.java
@@ -4,27 +4,27 @@ import net.minecraft.nbt.NbtCompound;
 import net.minecraft.world.PersistentState;
 
 public class InventoryWeightState extends PersistentState {
-    private long maxWeight = 10000; // Default value
+    private float maxWeight = 10000; // Default value
 
-    public void setMaxWeight(long value) {
+    public void setMaxWeight(float value) {
         this.maxWeight = value;
         markDirty();
     }
 
-    public long getMaxWeight() {
+    public float getMaxWeight() {
         return this.maxWeight;
     }
 
     @Override
     public NbtCompound writeNbt(NbtCompound tag) {
-        tag.putLong("maxWeight", maxWeight);
+        tag.putFloat("maxWeight", maxWeight);
         return tag;
     }
 
     public static InventoryWeightState fromNbt(NbtCompound tag) {
         InventoryWeightState state = new InventoryWeightState();
         if (tag.contains("maxWeight")) {
-            state.maxWeight = tag.getLong("maxWeight");
+            state.maxWeight = tag.getFloat("maxWeight");
         }
         return state;
     }
